Add files via upload
The attached document summarizes the aspects of my realization of this project. My idea was to create the set of tables and data that encompass aspects of the activity of a football team within a season's timeframe. From ERD to tables' content, roles,  and functions-supported processes, everything is duly explained in the file!